golint/go vet changes to cleanup
