--- middlewares/auth.go
@@ -1,1 +1,0 @@
-package middlewares
